Mudando o local dos testes do CollectionManager
