+++ scielomanager/journalmanager/tests/tests_modelmanagers.py
@@ -1,5 +1,4 @@
 # coding: utf-8
-import unittest
 from django.test import TestCase
 from django_factory_boy import auth
 
@@ -8,6 +7,87 @@ from journalmanager import (
     modelmanagers,
 )
 from journalmanager.tests import modelfactories
+from scielomanager.utils.modelmanagers.helpers import (
+    _makeUserProfile,
+    _makeUserRequestContext,
+)
+
+
+class CollectionManagerTests(TestCase):
+    def setUp(self):
+        self.user = auth.UserF()
+        _makeUserProfile(self.user)
+        _makeUserRequestContext(self.user)
+        self.collection = modelfactories.CollectionFactory.create()
+
+    def tearDown(self):
+        """
+        Restore the default values.
+        """
+
+    def test_get_all_by_user(self):
+        for i in range(5):
+            if i % 2:
+                modelfactories.CollectionFactory.create()
+            else:
+                col = modelfactories.CollectionFactory.create()
+                col.add_user(self.user)
+
+        def get_user_collections():
+            return self.user.user_collection.all()
+
+        collections = models.Collection.userobjects.all(
+            get_all_collections=get_user_collections
+        )
+
+        self.assertEqual(collections.count(), 3)
+
+    def test_get_default_by_user(self):
+        col1 = modelfactories.CollectionFactory.create()
+        col1.make_default_to_user(self.user)
+        col2 = modelfactories.CollectionFactory.create()
+        col2.add_user(self.user)
+        self.assertEqual(
+            self.user.get_profile().get_default_collection,
+            col1)
+
+    def test_get_default_by_user_second_collection(self):
+        col1 = modelfactories.CollectionFactory.create()
+        col1.make_default_to_user(self.user)
+        col2 = modelfactories.CollectionFactory.create()
+        col2.make_default_to_user(self.user)
+
+        def get_user_active_collection():
+            return col2
+
+        self.assertEqual(
+            models.Collection.userobjects.active(get_active_collection=get_user_active_collection),
+            col2)
+
+    def test_get_default_by_user_with_two_users(self):
+        user1 = auth.UserF()
+        user2 = auth.UserF()
+        _makeUserProfile(user1)
+        _makeUserProfile(user2)
+
+        col1 = modelfactories.CollectionFactory.create()
+        col1.make_default_to_user(user1)
+
+        col2 = modelfactories.CollectionFactory.create()
+        col2.add_user(user1)
+        col2.make_default_to_user(user2)
+        self.assertEqual(user1.get_profile().get_default_collection, col1)
+        self.assertEqual(user2.get_profile().get_default_collection, col2)
+
+    def test_get_default_by_user_must_raise_doesnotexist_if_the_user_has_no_collections(self):
+        col1 = modelfactories.CollectionFactory.create()
+
+        def get_user_collections():
+            raise RuntimeError()
+
+        self.assertRaises(
+            models.Collection.DoesNotExist,
+            lambda: models.Collection.userobjects.active(get_active_collection=get_user_collections))
 
 
 class JournalManagerTests(TestCase):
